Removing multiple db initialization and moving create 411 handler to the nostr handlers folder
