+++ lib/web/handler_login.go
@@ -5,11 +5,13 @@ import (
 	"time"
 
 	"golang.org/x/crypto/bcrypt"
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 
 	types "github.com/HORNET-Storage/hornet-storage/lib"
-	"github.com/HORNET-Storage/hornet-storage/lib/stores/graviton"
 	"github.com/gofiber/fiber/v2"
 	"github.com/nbd-wtf/go-nostr"
+	"github.com/spf13/viper"
 )
 
 func loginUser(c *fiber.Ctx) error {
@@ -23,12 +25,16 @@ func loginUser(c *fiber.Ctx) error {
 		})
 	}
 
-	db, err := graviton.InitGorm()
+	dbPath := viper.GetString("relay_stats_db")
+	if dbPath == "" {
+		log.Fatal("Database path not found in config")
+	}
+
+	// Initialize the Gorm database
+	db, err := gorm.Open(sqlite.Open(dbPath), &gorm.Config{})
 	if err != nil {
 		log.Printf("Failed to connect to the database: %v", err)
-		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"error": "Internal server error",
-		})
+		return c.Status(fiber.StatusInternalServerError).SendString("Internal Server Error")
 	}
 
 	var user types.User
