+++ lib/web/handler_logout.go
@@ -5,8 +5,10 @@ import (
 	"strings"
 
 	types "github.com/HORNET-Storage/hornet-storage/lib"
-	"github.com/HORNET-Storage/hornet-storage/lib/stores/graviton"
 	"github.com/gofiber/fiber/v2"
+	"github.com/spf13/viper"
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 )
 
 func logoutUser(c *fiber.Ctx) error {
@@ -21,12 +23,16 @@ func logoutUser(c *fiber.Ctx) error {
 	token = strings.TrimPrefix(token, "Bearer ")
 
 	// Initialize the database connection
-	db, err := graviton.InitGorm()
+	dbPath := viper.GetString("relay_stats_db")
+	if dbPath == "" {
+		log.Fatal("Database path not found in config")
+	}
+
+	// Initialize the Gorm database
+	db, err := gorm.Open(sqlite.Open(dbPath), &gorm.Config{})
 	if err != nil {
 		log.Printf("Failed to connect to the database: %v", err)
-		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
-			"error": "Internal server error",
-		})
+		return c.Status(fiber.StatusInternalServerError).SendString("Internal Server Error")
 	}
 
 	// Delete the token from ActiveTokens
