+++ lib/web/handler_pull_wallet_addresses.go
@@ -4,8 +4,10 @@ import (
 	"log"
 
 	"github.com/HORNET-Storage/hornet-storage/lib"
-	"github.com/HORNET-Storage/hornet-storage/lib/stores/graviton"
 	"github.com/gofiber/fiber/v2"
+	"github.com/spf13/viper"
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 )
 
 // AddressResponse represents the format of the address data to be returned
@@ -18,7 +20,13 @@ func pullWalletAddresses(c *fiber.Ctx) error {
 	log.Println("Get addresses request received")
 
 	// Initialize the Gorm database
-	db, err := graviton.InitGorm()
+	dbPath := viper.GetString("relay_stats_db")
+	if dbPath == "" {
+		log.Fatal("Database path not found in config")
+	}
+
+	// Initialize the Gorm database
+	db, err := gorm.Open(sqlite.Open(dbPath), &gorm.Config{})
 	if err != nil {
 		log.Printf("Failed to connect to the database: %v", err)
 		return c.Status(fiber.StatusInternalServerError).SendString("Internal Server Error")
