+++ lib/web/handler_wallet_balance.go
@@ -5,14 +5,21 @@ import (
 	"strconv"
 
 	types "github.com/HORNET-Storage/hornet-storage/lib"
-	"github.com/HORNET-Storage/hornet-storage/lib/stores/graviton"
 	"github.com/gofiber/fiber/v2"
+	"github.com/spf13/viper"
+	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 )
 
 func getWalletBalanceUSD(c *fiber.Ctx) error {
 	// Initialize the Gorm database
-	db, err := graviton.InitGorm()
+	dbPath := viper.GetString("relay_stats_db")
+	if dbPath == "" {
+		log.Fatal("Database path not found in config")
+	}
+
+	// Initialize the Gorm database
+	db, err := gorm.Open(sqlite.Open(dbPath), &gorm.Config{})
 	if err != nil {
 		log.Printf("Failed to connect to the database: %v", err)
 		return c.Status(fiber.StatusInternalServerError).SendString("Internal Server Error")
